# src/Geolocation.Client/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Geolocation.App",
    "version": "v1"
  },
  "paths": {
    "/Addresses": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Список адресов.",
        "operationId": "AddressesGET",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/IAddress"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Addresses"
        ],
        "summary": "Список адресов.",
        "operationId": "AddressesPOST",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddressDto"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Geolocation": {
      "get": {
        "tags": [
          "Geolocation"
        ],
        "summary": "Возвращает всю информацию об адресе по координатам. Работает для домов, улиц и городов.",
        "operationId": "GeolocationGET",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "description": "Географическая широта",
            "type": "number",
            "format": "double"
          },
          {
            "in": "query",
            "name": "lon",
            "description": "Географическая долгота",
            "type": "number",
            "format": "double"
          }
        ],
        "responses": {
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AddressDto"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "AddressDto": {
      "required": [
        "value"
      ],
      "type": "object",
      "properties": {
        "correlationId": {
          "format": "uuid",
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "unrestrictedValue": {
          "type": "string"
        },
        "postalCode": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "kladrId": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/AddressType"
        }
      },
      "example": "{\n  \"CorrelationId\": \"92f522d0-b88c-4299-a7d0-43b70600f939\",\n  \"Value\": \"Белгородская обл, г Старый Оскол, мкр Дубрава квартал 3\",\n  \"UnrestrictedValue\": \"309502, Белгородская обл, г Старый Оскол, мкр Дубрава квартал 3\",\n  \"PostalCode\": \"309502\",\n  \"Country\": \"Россия\",\n  \"KladrId\": \"3100000200004070007\",\n  \"Type\": \"Work\"\n}"
    },
    "AddressType": {
      "description": "Тип адреса",
      "enum": [
        "Registration",
        "Home",
        "Work"
      ],
      "type": "string"
    },
    "IAddress": {
      "type": "object",
      "properties": {
        "correlationId": {
          "format": "uuid",
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "unrestrictedValue": {
          "type": "string"
        },
        "postalCode": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "kladrId": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/AddressType"
        }
      },
      "example": "{\n  \"CorrelationId\": \"92f522d0-b88c-4299-a7d0-43b70600f939\",\n  \"Value\": \"Белгородская обл, г Старый Оскол, мкр Дубрава квартал 3\",\n  \"UnrestrictedValue\": \"309502, Белгородская обл, г Старый Оскол, мкр Дубрава квартал 3\",\n  \"PostalCode\": \"309502\",\n  \"Country\": \"Россия\",\n  \"KladrId\": \"3100000200004070007\",\n  \"Type\": \"Work\"\n}"
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": {},
      "example": "{\n  \"Type\": \"Microsoft.AspNetCore.Http.BadHttpRequestException\",\n  \"Title\": \"One or more validation errors occurred\",\n  \"Status\": 400,\n  \"Extensions\": {}\n}"
    }
  }
}